moved out jointree transformations into its own file
