+++ go/vt/vtgate/planbuilder/route_planning.go
@@ -18,9 +18,6 @@ package planbuilder
 
 import (
 	"sort"
-	"strings"
-
-	"vitess.io/vitess/go/sqltypes"
 
 	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/vtgate/semantics"
@@ -128,6 +125,10 @@ type (
 		// vindex and conditions is set if a vindex will be used for this route.
 		vindex     vindexes.Vindex
 		conditions []sqlparser.Expr
+
+		// here we store the possible vindexes we can use so that when we add predicates to the plan,
+		// we can quickly check if the new predicates enables any new vindex options
+		vindexPreds []vindexPlusPredicates
 	}
 	joinPlan struct {
 		predicates []sqlparser.Expr
@@ -136,7 +137,16 @@ type (
 	routeTables []*routeTable
 )
 
-// solves implements the joinTree interface
+// clone returns a copy of the struct with copies of slices,
+// so changing the the contents of them will not be reflected in the original
+func (rp *routePlan) clone() *routePlan {
+	result := *rp
+	result.vindexPreds = make([]vindexPlusPredicates, len(rp.vindexPreds))
+	copy(result.vindexPreds, rp.vindexPreds)
+	return &result
+}
+
+// solves returns the tables that this plan is solving
 func (rp *routePlan) solves() semantics.TableSet {
 	return rp.solved
 }
@@ -149,8 +159,9 @@ func (*routePlan) cost() int {
 // vindexPlusPredicates is a struct used to store all the predicates that the vindex can be used to query
 type vindexPlusPredicates struct {
 	vindex     *vindexes.ColumnVindex
-	covered    bool
 	predicates []sqlparser.Expr
+	// Vindex is covered if all the columns in the vindex have an associated predicate
+	covered bool
 }
 
 func (rp *routePlan) addPredicate(predicates ...sqlparser.Expr) error {
@@ -414,142 +425,6 @@ func createRoutePlan(table *queryTable, solves semantics.TableSet, vschema Conte
 	}
 
 	return plan, nil
-}
-
-func transformToLogicalPlan(tree joinTree, semTable *semantics.SemTable) (logicalPlan, error) {
-	switch n := tree.(type) {
-	case *routePlan:
-		return transformRoutePlan(n)
-
-	case *joinPlan:
-		return transformJoinPlan(n, semTable)
-	}
-
-	return nil, vterrors.Errorf(vtrpcpb.Code_INTERNAL, "BUG: unknown type encountered: %T", tree)
-}
-
-func transformJoinPlan(n *joinPlan, semTable *semantics.SemTable) (*join2, error) {
-	lhsColList := extractColumnsNeededFromLHS(n, semTable, n.lhs.solves())
-
-	var lhsColExpr []*sqlparser.AliasedExpr
-	for _, col := range lhsColList {
-		lhsColExpr = append(lhsColExpr, &sqlparser.AliasedExpr{
-			Expr: col,
-		})
-	}
-
-	lhs, err := transformToLogicalPlan(n.lhs, semTable)
-	if err != nil {
-		return nil, err
-	}
-	offset, err := pushProjection(lhsColExpr, lhs, semTable)
-	if err != nil {
-		return nil, err
-	}
-
-	vars := map[string]int{}
-
-	for _, col := range lhsColList {
-		vars[col.CompliantName("")] = offset
-		offset++
-	}
-
-	rhs, err := transformToLogicalPlan(n.rhs, semTable)
-	if err != nil {
-		return nil, err
-	}
-
-	err = pushPredicate(n.predicates, rhs, semTable)
-	if err != nil {
-		return nil, err
-	}
-
-	return &join2{
-		Left:  lhs,
-		Right: rhs,
-		Vars:  vars,
-	}, nil
-}
-
-func extractColumnsNeededFromLHS(n *joinPlan, semTable *semantics.SemTable, lhsSolves semantics.TableSet) []*sqlparser.ColName {
-	lhsColMap := map[*sqlparser.ColName]sqlparser.Argument{}
-	for _, predicate := range n.predicates {
-		sqlparser.Rewrite(predicate, func(cursor *sqlparser.Cursor) bool {
-			switch node := cursor.Node().(type) {
-			case *sqlparser.ColName:
-				if semTable.Dependencies(node).IsSolvedBy(lhsSolves) {
-					arg := sqlparser.NewArgument([]byte(":" + node.CompliantName("")))
-					lhsColMap[node] = arg
-					cursor.Replace(arg)
-				}
-			}
-			return true
-		}, nil)
-	}
-
-	var lhsColList []*sqlparser.ColName
-	for col := range lhsColMap {
-		lhsColList = append(lhsColList, col)
-	}
-	return lhsColList
-}
-
-func transformRoutePlan(n *routePlan) (*route, error) {
-	var tablesForSelect sqlparser.TableExprs
-	tableNameMap := map[string]interface{}{}
-
-	sort.Sort(n.tables)
-	for _, t := range n.tables {
-		alias := sqlparser.AliasedTableExpr{
-			Expr: sqlparser.TableName{
-				Name: t.vtable.Name,
-			},
-			Partitions: nil,
-			As:         t.qtable.alias.As,
-			Hints:      nil,
-		}
-		tablesForSelect = append(tablesForSelect, &alias)
-		tableNameMap[sqlparser.String(t.qtable.table.Name)] = nil
-	}
-
-	predicates := n.Predicates()
-	var where *sqlparser.Where
-	if predicates != nil {
-		where = &sqlparser.Where{Expr: predicates, Type: sqlparser.WhereClause}
-	}
-	var values []sqltypes.PlanValue
-	if len(n.conditions) == 1 {
-		value, err := sqlparser.NewPlanValue(n.conditions[0].(*sqlparser.ComparisonExpr).Right)
-		if err != nil {
-			return nil, err
-		}
-		values = []sqltypes.PlanValue{value}
-	}
-	var singleColumn vindexes.SingleColumn
-	if n.vindex != nil {
-		singleColumn = n.vindex.(vindexes.SingleColumn)
-	}
-
-	var tableNames []string
-	for name := range tableNameMap {
-		tableNames = append(tableNames, name)
-	}
-	sort.Strings(tableNames)
-
-	return &route{
-		eroute: &engine.Route{
-			Opcode:    n.routeOpCode,
-			TableName: strings.Join(tableNames, ", "),
-			Keyspace:  n.keyspace,
-			Vindex:    singleColumn,
-			Values:    values,
-		},
-		Select: &sqlparser.Select{
-			From:  tablesForSelect,
-			Where: where,
-		},
-		solvedTables: n.solved,
-	}, nil
 }
 
 func findColumnVindex(a *routePlan, exp sqlparser.Expr, sem *semantics.SemTable) vindexes.SingleColumn {
